% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-rhocap-check-dot.R
\name{.check_rhocap}
\alias{.check_rhocap}
\title{Sanity Check for the Correlation Matrix}
\usage{
.check_rhocap(x, return_k = FALSE)
}
\arguments{
\item{x}{Numeric matrix.
Correlation matrix.}

\item{return_k}{Logical.
Return the dimension of the covariance matrix.}
}
\description{
Sanity Check for the Correlation Matrix
}
\seealso{
Other Correlation Functions: 
\code{\link{.check_vechsrhocap}()},
\code{\link{.rhocap_of_sigmacap}()},
\code{\link{.sigmacap_of_rhocap}()},
\code{\link{cor_of_cov}()},
\code{\link{cov_of_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{check}
\keyword{correlation}
\keyword{rhoMatrix}
