% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-vechsrhocap-check-dot.R
\name{.check_vechsrhocap}
\alias{.check_vechsrhocap}
\title{Sanity Check for the Strict Half-Vectorization
of the Correlation Matrix}
\usage{
.check_vechsrhocap(x, return_k = FALSE)
}
\arguments{
\item{x}{Numeric vector.
Strict half-vectorization of the covariance matrix.}

\item{return_k}{Logical.
Return the dimension of the covariance matrix.}
}
\description{
Sanity Check for the Strict Half-Vectorization
of the Correlation Matrix
}
\seealso{
Other Correlation Functions: 
\code{\link{.check_rhocap}()},
\code{\link{.rhocap_of_sigmacap}()},
\code{\link{.sigmacap_of_rhocap}()},
\code{\link{cor_of_cov}()},
\code{\link{cov_of_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{check}
\keyword{correlation}
\keyword{rhoMatrix}
