% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-rhocap_of_sigmacap-dot.R
\name{.rhocap_of_sigmacap}
\alias{.rhocap_of_sigmacap}
\title{Correlation Matrix from Covariance Matrix}
\usage{
.rhocap_of_sigmacap(x, q)
}
\arguments{
\item{x}{Numeric matrix.
Covariance matrix.}

\item{q}{Numeric vector.
Inverse of the standard deviation vector.}
}
\description{
Correlation Matrix from Covariance Matrix
}
\seealso{
Other Correlation Functions: 
\code{\link{.check_rhocap}()},
\code{\link{.check_vechsrhocap}()},
\code{\link{.sigmacap_of_rhocap}()},
\code{\link{cor_of_cov}()},
\code{\link{cov_of_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{correlation}
\keyword{dot}
\keyword{rhoMatrix}
