% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-sigmacap_of_rhocap-dot.R
\name{.sigmacap_of_rhocap}
\alias{.sigmacap_of_rhocap}
\title{Covariance Matrix from Correlation Matrix}
\usage{
.sigmacap_of_rhocap(x, sd)
}
\arguments{
\item{x}{Numeric matrix.
Correlation matrix.}

\item{sd}{Numeric vector.
Vector of standard deviations.}
}
\description{
Covariance Matrix from Correlation Matrix
}
\seealso{
Other Correlation Functions: 
\code{\link{.check_rhocap}()},
\code{\link{.check_vechsrhocap}()},
\code{\link{.rhocap_of_sigmacap}()},
\code{\link{cor_of_cov}()},
\code{\link{cov_of_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{correlation}
\keyword{dot}
\keyword{rhoMatrix}
