% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-cor_of_cov.R,
%   R/rhoMatrix-rhocap_of_sigmacap.R
\name{cor_of_cov}
\alias{cor_of_cov}
\alias{rhocap_of_sigmacap}
\title{Correlation Matrix from Covariance Matrix}
\usage{
cor_of_cov(x)

rhocap_of_sigmacap(x)
}
\arguments{
\item{x}{Numeric matrix.
Covariance matrix.}
}
\description{
Correlation Matrix from Covariance Matrix
}
\seealso{
Other Correlation Functions: 
\code{\link{cov_of_cor}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{correlation}
\keyword{rhoMatrix}
