% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-package.R
\docType{package}
\name{rhoMatrix}
\alias{rhoMatrix}
\title{rhoMatrix: Correlation Matrix}
\description{
A collection of functions related
to the correlation matrix \eqn{\mathbf{P}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{package}
\keyword{rhoMatrix}
