% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-cov_of_cor.R,
%   R/rhoMatrix-sigmacap_of_rhocap.R
\name{cov_of_cor}
\alias{cov_of_cor}
\alias{sigmacap_of_rhocap}
\title{Covariance Matrix from Correlation Matrix}
\usage{
cov_of_cor(x, sd)

sigmacap_of_rhocap(x, sd)
}
\arguments{
\item{x}{Numeric matrix.
Correlation matrix.}

\item{sd}{Numeric vector.
Vector of standard deviations.}
}
\description{
Covariance Matrix from Correlation Matrix
}
\seealso{
Other Correlation Functions: 
\code{\link{cor_of_cov}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{correlation}
\keyword{rhoMatrix}
